fix(issue-29): panic when different pan IDs
When the PAN IDs are equal, then PAN ID compression is used, removing
the source PAN ID from the frame. The same behaviour was observed when
the PAN IDs were different, which is incorrect. This commit fixes the
issue by only compressing the source PAN ID when the PAN IDs are equal.